Handle io errors while reading by reconnecting.
This may not be correct. I don't quite know how to introspect the errors
and discern between a temporary problem and one that is permanent for
the connection. But it won't be too hard to do that in the error handler
where right now we just always send to the reconnect channel.