Add test for connectionManager and pollingManager
This required refactoring a lot, because self and the implicit rules
around it seemed too limiting. This meant channels need to be created
outside of the objects, but that may be easily wrappable in a generic
"start things up" function later. It also seems that the 'running' flag
may be unnecessary if we just use the stop channel instead, but that can
be worked out in later revisions.