Add a select_connection method
This method will be used by clients to send a similar ID to the same
server to make coalescing more effective. Note that the test panics
because we haven't added connections. We'll add a wait-for-connection
method and then add servers and flip that off once we have a pattern
for testing a real server.