Major refactoring to add logging and wait_for_conn
The method wait_for_connection and any others that deal directly
with locks likely need to be moved ot a higher level struct at
some time so we can free the user from managing locks and just allow
them to write gearman related calls. In this case the method would never
work as it was, because read locking the structure and holding it open to
wait on the condition variable prevented the other threads from write
locking the structure to update the connection list and trigger said
condition variable, leading to a deadlock.