Fixing logic in queues get_job
This is still likely broken in other ways, but basically because of the
WrappingHashSet, we weren't actually finding workers to put to sleep
when getting PRE_SLEEP's. These two additional scripts still don't pass,
because we don't support _BG jobs. But they are intended to test the
corner case of a worker that registers multiple functions.