Implement coalescing that works
Now two clients submitting foreground jobs with the same unique will
only cause one job to be pushed into the system, and only one worker
will do the work, and the reply will go to both clients.

Unfortunately, because of this we have to use Rc's since we have two
things pointing to our jobs: a hash map by unique, and a job queue.

Originally I'd wanted to use Weak's for the queue, to allow for a future
job cancellation system. However, somehow the refcount was dipping to 0
and the jobs were being dropped, despite the Rc's being stored in a
hashmap. For now, it's ok. Even if a job is cancelled (something we have
no facility for) it will still be sent to a worker. We can revisit
Weak's for this later.