Tests with workers, Client.do_one_job added
We can now do tests that complete jobs. In creating the first test of
this kind, it became clear there's no need to call the work() method if
we just have a do_one_job method. We'll keep work() for ergonomic
reasons as it makes workers very easy to create.