Fix background job response errors
We promise in the API that background jobs will return an error on call
to response, but we were blocking forever because the tx was present
even though it would never receive the WORK_* update packets from the
server. This test also will verify that behavior, though unfortunately
we have to poll unless we want to make a worker that communicates via a
channel back to the test which seems like overkill.