Test that hash ring is routing properly
We expect that two of our clients who get the same uniq id in submit
will send a job to the same server, increasing the odds that work will
be reduced.

In running this test over and over with logging enabled I saw some
explosions around frame reading, basically where we saw the connection
teardown interfere with reading from the socket. This is normal and we
expect that this connection drop will be handled after this error, but
just in case it's happening a lot a warn seems appropriate.