Add server fallback via hashring to client
This makes gearman more resilient, and is how it was originally intended
to work, though clients vary in their adherence to the behavior. We
don't have a way to configure the # of replicas yet but it's a goal for
the future.

Rather than make another test that starts 10 servers, instead we just
tack tests for this behavior onto the end, as making the servers is
pretty slow.