Fixing PRE_SLEEP handling in server
This started as a test for coalescing behavior.

Testing for not-coalescing showed some problems with the way PRE_SLEEP
was handled. If there are jobs for a worker that has sent PRE_SLEEP, we
should wake it up immediately, not wait for another job to bump the
queue.

And doing that exposed a bigger problem with WrappingHashset: In
versions prior to 0.5, two different iterators would interfere with each
other. So the library was updated to have iterators keep the count, and
the Wrappinghashset object keep the position, which works in a more
expected way, but is an API break. Without the upgrade though, has_job
causes the next get_job to return None and this puts the client and
server into a death spiral.