Addressing lock panics in Client
We actually should keep the connections locked until we send the packet
so the list doesn't change, as that will avoid the potential for a panic
while holding the lock.

We are changing all of the unwraps here to expects as we have reviewed
all of the execution paths from the lock and they don't have known
panics.

This is but one of many changes needed to address issue #4