Support multi-address hosts in client
Previously the connector mechanism would always just take the first
address and try it. If somebody wants to use rrdns that won't work
properly. This will try all of the addresses that a server has before
giving up on the server for the RETRY backoff.

This also removes an unwrap, which is a step toward addressing issue#4.