BREAKING: Refactor error handling
Errors have been largely returned as io::Error for everything up until
now. But there are other errors which are panics now, that should be
results. Rather than bend over backward implementing froms for all of
them, we'll make the errors dynamic.

This is expected to break client code if client code was set up to
inspect the errors returning. As such, we are also changing the
ErrorKind for some of the io::Errors we're returning.