Add low-level packet sending to WorkerJob
This will allow workers to construct any Gearman packet they need to
send, which can be a little dangerous but may be useful for
compatibility with other clients. We specifically use it here to send a
broken packet to test our client's handling of it.

With this we re-work some error handling to make the client more
resilient, and test that it is, in fact, more resilient.