Better server error support
Rustygeard was unwrapping and panicing on basic string problems and
generally not sending good errors. There are more of these, but this
starts the process of both being able to send them, and receive them in
the client. The server error codes are not defined in the Gearman
protocol, they seem to be server-specific, so we needed to define a few
here.